Moved DPI related CLI functions into a specific file. (#22)
