+++ src/plugins/upf/upf.h
@@ -29,8 +29,6 @@
 #include <vppinfra/bihash_24_8.h>
 
 #include "pfcp.h"
-
-#include "dpi.h"
 #include "flowtable_impl.h"
 #include "flowtable.h"
 
@@ -446,6 +444,8 @@ typedef struct {
   pfcp_recovery_time_stamp_t recovery_time_stamp;
 } upf_node_assoc_t;
 
+typedef u8 * regex_t;
+
 typedef struct {
   u32 id;
   regex_t host;
@@ -541,12 +541,6 @@ int upf_enable_disable (upf_main_t * sm, u32 sw_if_index,
 			  int enable_disable);
 u8 * format_upf_encap_trace (u8 * s, va_list * args);
 void gtpu_send_end_marker(upf_far_forward_t * forward);
-
-int upf_app_add_del (upf_main_t * sm, u8 * name, int add);
-int upf_rule_add_del (upf_main_t * sm, u8 * name, u32 id,
-                      int add, upf_rule_args_t * args);
-
-int upf_add_multi_regex(u8 ** apps, u32 * db_index, u8 create);
 
 #endif /* __included_upf_h__ */
