+++ src/plugins/upf/upf_api.c
@@ -28,6 +28,7 @@
 #include <vlibmemory/api.h>
 
 #include <upf/upf.h>
+#include <upf/dpi.h>
 
 /* define message IDs */
 #include <upf/upf_msg_enum.h>
